refactor(goesr): move DatasetMetadata class
